added new file structure
