+++ pkg/oauth_provider/github/provider.go
@@ -1,0 +1,1 @@
+package github
