+++ pkg/oauth_provider/google/provider.go
@@ -1,0 +1,1 @@
+package google
