+++ pkg/session_repository/inmemory/repository.go
@@ -1,0 +1,1 @@
+package inmemeory
