+++ pkg/shortener/link_repository.go
@@ -1,0 +1,1 @@
+package shortener
