+++ pkg/shortener/oauth_provider.go
@@ -1,0 +1,1 @@
+package shortener
