--- users/user_service_test.go
@@ -1,79 +1,0 @@
-package users
-
-import (
-	"log"
-	"math/rand"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-var testStore = &Database{}
-
-func init() {
-	testStore = InitializeDatabase()
-}
-
-func TestStoreInit(t *testing.T) {
-	assert.NotNil(t, testStore.db)
-}
-
-func TestGetUser(t *testing.T) {
-	id := "guest"
-	retid := GetUser(id)
-	assert.Equal(t, id, retid)
-}
-
-func TestCreateUser(t *testing.T) {
-	randomString := func (length int) string {
-		const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-		b := make([]byte, length)
-		for i := range b {
-			b[i] = letterBytes[rand.Intn(len(letterBytes))]
-		}
-		return string(b)
-	} 
-	id := randomString(8);
-	err := CreateUser(id, "");
-	if err != nil {
-		log.Fatal(err)
-	}
-	retid := GetUser(id)
-	assert.Equal(t, retid, id)
-	deleteUser(id)
-}
-
-
-func TestAddAndGetLinksByUser(t *testing.T) {
-	randomString := func (length int) string {
-		const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-		b := make([]byte, length)
-		for i := range b {
-			b[i] = letterBytes[rand.Intn(len(letterBytes))]
-		}
-		return string(b)
-	} 
-	id := randomString(8)
-	err := CreateUser(id, "")
-	if err != nil {
-		log.Fatal(err)
-	}
-	AddLink(id, "one", "two", "his");
-	links := GetLinksByUser(id)
-	flag := false
-	for _, link := range links {
-		if link.short == "one" && link.long == "two" {
-			flag = true
-		}
-	}
-	assert.Equal(t, flag, true)
-	deleteUserAndLinks(id)
-}
-
-func TestDeleteLink(t *testing.T) {
-	info := GetLinksByUser("guest")
-	AddLink("guest", "hi", "bye", "unique_id")
-	DeleteLink("unique_id")
-	info2 := GetLinksByUser("guest")
-	assert.Equal(t, len(info), len(info2))
-}
-
